requirements/base.txt
argon2-cffi
redis

# Django
# ------------------------------------------------------------------------------
django
django-environ
django-model-utils
django-crispy-forms
django-tables2
django-filter
django-redis

# Django REST Framework
djangorestframework
coreapi

channels
channels-redis
